Final warning fixes - batch 8 (ZERO WARNINGS TARGET)
Fixed remaining 6 warnings:
- zeta-graph-kv-integration.h: Mark unused salience_threshold parameter
- zeta-hrm.h: Mark unused response parameter
- zeta-output-control.h: Fix sign comparison with explicit casts
- zeta-scratch-integration.h: Mark unused array with __attribute__((unused))
- zeta-tools.h: Add (void)value to suppress unused param warning
- zeta-trm.h: Mark unused query parameter

ZERO WARNINGS = ENGINEER RESPECT 🎯✨

+++ llama.cpp/tools/zeta-zero/zeta-output-control.h
@@ -460,10 +460,10 @@ static inline std::string zeta_postprocess_output(const std::string& output,
     std::string result = output;
 
     // Enforce character limit
-    if ((int)result.size() > ctrl.max_chars) {
+    if (result.size() > (size_t)ctrl.max_chars) {
         // Find a good break point (end of sentence)
-        size_t break_pos = ctrl.max_chars;
-        for (size_t i = ctrl.max_chars; i > ctrl.max_chars - 200 && i > 0; i--) {
+        size_t break_pos = (size_t)ctrl.max_chars;
+        for (size_t i = (size_t)ctrl.max_chars; i > (size_t)ctrl.max_chars - 200 && i > 0; i--) {
             if (result[i] == '.' || result[i] == '!' || result[i] == '?' ||
                 result[i] == '\n' || result[i] == '}') {
                 break_pos = i + 1;
